refactor: watcher for column picker
